Move lifecycle to new separate ix_apps package
